requirements.txt
Django==3.2.8  # https://www.djangoproject.com/
djangorestframework==3.12.4  # https://github.com/encode/django-rest-framework

amqp==5.0.9
asttokens==2.0.5
backcall==0.2.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.1
decorator==5.1.1
django-filter==21.1
django-redis==5.2.0
django-redis-cache==3.0.0
djangorestframework-jsonapi==5.0.0
executing==0.8.2
idna==3.3
ipdb==0.13.9
ipython==8.1.0
jedi==0.18.1
kombu==5.2.3
PyAMQP==0.1.0.7
matplotlib-inline==0.1.3
msgpack==1.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.28
psycopg2-binary>=2.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
redis==3.5.3
requests==2.27.1
six==1.16.0
stack-data==0.2.0
toml==0.10.2
traitlets==5.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5